Extract client flags specs to seperate methods
Make sure the main function is not overhelming,
extract portion of lines to separate methods for better read.

Found that client in `toxic add` supports both: `--upstream` and
`--downstream`. Because it tries to create 2 toxic with single name it
produces the error when tries add `--downstream` toxic, and in same time
creates the `--upstream` toxic.

To remove confusion update the client usage help to support only one in
same time. Return error message in case 2 flags provided.