Extract client flags specs to seperate methods (#321)
Make sure the main function in client is not overwhelming,
extract portion of lines to separate methods for better read.

Found that client in `toxic add` supports both: `--upstream` and
`--downstream`. Because it tries to create 2 toxics with same name it
produces the error when tries add `--downstream` toxic, and in same time
creates the `--upstream` toxic.

To remove confusion update the client usage help to support only one in
same time. Return error message in case 2 flags provided.

Introduce a new way to add toxics with client: 
Provide a structure`ToxicOptions` with all required options and pass to `client.AddToxic`, `client.UpdateToxic` and `client.RemoveToxic`