Handle slicer Toxic with zero SizeVariation (#359)
This change skips randomized size variation for the slicer toxic if the SizeVariation parameter is 0, which is also the default setting.

Additionally, it removes `rand.Intn(1)` was being added to mid to adjust for integer division rounding. However, rand.Intn(n) returns an integer in the range [0, n), so rand.Intn(1) always returns 0. It was removed.