Update README.md - clarify timeout toxic behavior
The code for the timeout toxic when timeout is set to zero simply reads from the input and then if it gets some data just has a comment: "// Drop the data on the ground".  This change updates the docs do indicate the data is dropped, not delayed, until the texic is removed.