Bugfix: don't stop proxies when using hostnames to specify listen addresses
Before this change we'd stop existing proxies when populating a set of proxies
when the listen addresses were specified using hostnames instead of IP
addresses.

This is very surprising as we encourage users to populate proxies at the begin
of an application start:
> A /populate call can be included for example at application start to ensure
> all required proxies exist. It is safe to make this call several times, since
> proxies will be untouched as long as their fields are consistent with the new
> data.

This is (currently) only partially true as it doesn't work when using hostnames.

For upstream services this isn't a problem because we never store the resolved address.

The change introduces a "Differs" function on a proxy to determine if two
proxies are different in terms of listening address and upstream service. This
is now used both when changing collections as well as updating individual
proxies.

Closes https://github.com/Shopify/toxiproxy/issues/131.